Fixed types in crop and resize_crop_mirror tests

+++ dali/pipeline/operators/fused/resize_crop_mirror_test.cc
@@ -39,27 +39,27 @@ TYPED_TEST_CASE(ResizeCropMirrorTest, Types);
 TYPED_TEST(ResizeCropMirrorTest, TestFixedResizeAndCrop) {
   this->TstBody(this->DefaultSchema()
                 .AddArg("resize_shorter", 480.f)
-                .AddArg("crop", vector<int>{224, 224}), 5e-6);
+                .AddArg("crop", vector<float>{224, 224}), 5e-6);
 }
 
 TYPED_TEST(ResizeCropMirrorTest, TestFixedResizeAndCropWarp) {
   this->TstBody(this->DefaultSchema()
                 .AddArg("resize_x", 480.f)
                 .AddArg("resize_y", 480.f)
-                .AddArg("crop", vector<int>{224, 224}), 5e-6);
+                .AddArg("crop", vector<float>{224, 224}), 5e-6);
 }
 
 TYPED_TEST(ResizeCropMirrorTest, TestFixedFastResizeAndCrop) {
   this->TstBody(this->DefaultSchema(true)
                 .AddArg("resize_shorter", 480.f)
-                .AddArg("crop", vector<int>{224, 224}), 1.98);
+                .AddArg("crop", vector<float>{224, 224}), 1.98);
 }
 
 TYPED_TEST(ResizeCropMirrorTest, TestFixedFastResizeAndCropWarp) {
   this->TstBody(this->DefaultSchema(true)
                 .AddArg("resize_x", 480.f)
                 .AddArg("resize_y", 480.f)
-                .AddArg("crop", vector<int>{224, 224}), 1.80);
+                .AddArg("crop", vector<float>{224, 224}), 1.80);
 }
 
 }  // namespace dali
